+++ src/commands/unskip.ts
@@ -31,6 +31,7 @@ export default class implements Command {
         content: 'back \'er up\'',
         embeds: player.getCurrent() ? [buildPlayingMessageEmbed(player)] : [],
       });
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
     } catch (_: unknown) {
       throw new Error('no song to go back to');
     }