+++ src/utils/url.ts
@@ -14,6 +14,7 @@ export const cleanUrl = (url: string) => {
     }
 
     return u.toString();
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   } catch (_: unknown) {
     return url;
   }